pass correct -s and -m options to release tools
When building an OTA package, TARGET_RELEASETOOLS_EXTENSIONS can be
set (in BoardConfig.mk) to specify where the device-specific
releasetools code is located.  (The default location is the common
directory for the device's vendor.)  The TARGET_OTA_SCRIPT_MODE can be
used to override the default script mode ("auto") for a particular
device.

Conflicts:

	core/Makefile